to merge

+++ Aperture/Assets/Scripts/MapManager/MapManager.cs
@@ -44,7 +44,7 @@ public class MapManager : Singleton<MapManager>
 
 	public Vector2 GetTileCoord(Vector2 coord)
 	{
-		return new Vector2(coord.x / 0.16f, coord.y / 0.16f);
+		return new Vector2((int)(coord.x / 0.16f), (int)(coord.y / 0.16f));
 	}
 
 	void DisplayMap()
